requirements.txt
llama_index==0.4.28
openai==0.27.2
protobuf==4.22.1
requests==2.28.1
html2text==2020.1.16
google-cloud-storage
